Move all centers related components to the centers/ subfolder Put all container components in a containers/ subfolder Put all presentational components in a presentational/ subfolder
