+++ client/components/Main.jsx
@@ -3,8 +3,8 @@ import { Switch, Route, Redirect } from 'react-router-dom';
 import Home from './Home';
 import Login from './Login';
 import Signup from './Signup';
-import AddCenter from './AddCenter';
-import Centers from './Centers';
+import AddCenter from './centers/AddCenter';
+import Centers from './centers/Centers';
 import SingleCenter from './centers/SingleCenter';
 import EditCenter from './centers/EditCenter';
 import Events from './events/Events';
