+++ client/components/centers/containers/SingleCenter.jsx
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import EventsList from '../events/EventsList';
-import CenterDetail from './CenterDetail';
-import * as singleCenterActions from '../../actions/singleCenterActions';
+import EventsList from '../../events/EventsList';
+import CenterDetail from '../presentational/CenterDetail';
+import * as singleCenterActions from '../../../actions/singleCenterActions';
 
 class SingleCenter extends Component {
   // componentWillMount() {
